turned the contact page to a component, pending test
