+++ src/components/ContactForm.jsx
@@ -1,0 +1,98 @@
+import React, { useState } from 'react';
+import PropTypes from 'prop-types';
+
+const ContactForm = ({ onSubmit }) => {
+    const [formData, setFormData] = useState({
+        fullName: '',
+        email: '',
+        subject: '',
+        message: '',
+    });
+
+    const handleChange = (e) => {
+        const { name, value } = e.target;
+        setFormData((prevState) => ({
+            ...prevState,
+            [name]: value,
+        }));
+    };
+
+    const handleSubmit = (e) => {
+        e.preventDefault();
+        onSubmit(formData);
+    };
+
+    return (
+        <form onSubmit={handleSubmit} className="contact-form">
+            <div className="form-group">
+                <label htmlFor="fullName">
+                    Your Full Name <span className="required">*</span>
+                </label>
+                <input
+                    type="text"
+                    id="fullName"
+                    name="fullName"
+                    value={formData.fullName}
+                    onChange={handleChange}
+                    required
+                />
+            </div>
+
+            <div className="form-group">
+                <label htmlFor="email">
+                    Your Email <span className="required">*</span>
+                </label>
+                <input
+                    type="email"
+                    id="email"
+                    name="email"
+                    value={formData.email}
+                    onChange={handleChange}
+                    required
+                />
+            </div>
+
+            <div className="form-group">
+                <label htmlFor="subject">
+                    Subject <span className="required">*</span>
+                </label>
+                <select
+                    id="subject"
+                    name="subject"
+                    value={formData.subject}
+                    onChange={handleChange}
+                    required
+                >
+                    <option value="">Select...</option>
+                    <option value="Coaching Session">Coaching Session</option>
+                    <option value="Conference Request">Conference Request</option>
+                    <option value="Custom Request">Custom Request</option>
+                </select>
+            </div>
+
+            <div className="form-group">
+                <label htmlFor="message">
+                    Message <span className="required">*</span>
+                </label>
+                <textarea
+                    id="message"
+                    name="message"
+                    value={formData.message}
+                    onChange={handleChange}
+                    required
+                    rows={5}
+                ></textarea>
+            </div>
+
+            <button type="submit" className="submit-button">
+                Send Message
+            </button>
+        </form>
+    );
+};
+
+ContactForm.propTypes = {
+    onSubmit: PropTypes.func.isRequired,
+};
+
+export default ContactForm;
