--- src/pages/ContactPage.jsx
@@ -1,98 +1,0 @@
-import React, { useState } from 'react';
-import './ContactForm.css';
-
-const ContactForm = () => {
-  const [formData, setFormData] = useState({
-    fullName: '',
-    email: '',
-    subject: '',
-    message: '',
-  });
-
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    setFormData((prevState) => ({
-      ...prevState,
-      [name]: value,
-    }));
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    console.log(formData);
-    // Add your form submission logic here
-  };
-
-  return (
-    <div className="contact-form-phase1">
-      <div className="contact-role-model">
-        <h2>Contact Role Model</h2>
-        <p>Send a message to connect</p>
-        <form onSubmit={handleSubmit}>
-          <div className="form-group">
-            <label htmlFor="fullName">
-              Your Full Name <span className="required">*</span>
-            </label>
-            <input
-              type="text"
-              id="fullName"
-              name="fullName"
-              value={formData.fullName}
-              onChange={handleChange}
-              required
-            />
-          </div>
-
-          <div className="form-group">
-            <label htmlFor="email">
-              Your Email <span className="required">*</span>
-            </label>
-            <input
-              type="email"
-              id="email"
-              name="email"
-              value={formData.email}
-              onChange={handleChange}
-              required
-            />
-          </div>
-
-          <div className="form-group">
-            <label htmlFor="subject">
-              Subject <span className="required">*</span>
-            </label>
-            <select
-              id="subject"
-              name="subject"
-              value={formData.subject}
-              onChange={handleChange}
-              required
-            >
-              <option value="">Select...</option>
-            </select>
-          </div>
-
-          <div className="form-group">
-            <label htmlFor="message">
-              Message <span className="required">*</span>
-            </label>
-            <textarea
-              id="message"
-              name="message"
-              value={formData.message}
-              onChange={handleChange}
-              required
-              rows={5}
-            ></textarea>
-          </div>
-
-          <button type="submit" className="submit-button">
-            Send Message
-          </button>
-        </form>
-      </div>
-    </div>
-  );
-};
-
-export default ContactForm;
